requirements.txt
asgiref==3.7.2
coverage==7.3.1
Django==4.2.5
django-waffle==4.0.0
flake8==6.1.0
gitdb==4.0.10
GitPython==3.1.36
mccabe==0.7.0
pycodestyle==2.11.0
pyflakes==3.1.0
smmap==5.0.1
sqlparse==0.4.4
tzdata==2023.3
